test: Add a service upgrade test that verifies availability
The upstream upgrade test doesn't correctly test for disruption of
the backend during upgrade - it only tests that the service is
not available for more than 2 minutes at a time.

This is obviously not a good test of whether a service maintains
availability during upgrade. Take the core structure and use the
monitor and polling behavior of the apiserver which is much more
effective.